+++ CHANGELOG.md
@@ -1,11 +1,21 @@
+## [2025.2.2]
+
+### Fixes
+
+- Fixed [active project not detected at startup](https://github.com/hoxi/cmake-mux/issues/7)
+
+### Additions
+
+- Added [support remote development](https://github.com/hoxi/cmake-mux/issues/6)
+
 ## [2025.2.1]
 
-### Fixed
+### Fixes
 
 - Small tweak to make the preset loading more robust
 
 ## [2025.2]
 
-### Added
+### Additions
 
 - Initial release of plugin